The delete funtionality was protected by the user id.

+++ backend/routes/posts.js
@@ -68,10 +68,15 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     })
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post)
+
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
       .then(result => {
-        res.status(200).json({ message: "Update successful." })
+        if (result.nModified > 0) {
+          res.status(200).json({ message: "Update successful." })
+        }
+        else {
+          res.status(401).json({ message: "Not authorized!" })
+        }
       })
   })
 
@@ -115,9 +120,14 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      res.status(200).json({ message: "Post deleted." })
+      if (result.n > 0) {
+        res.status(200).json({ message: "Deletion successful." })
+      }
+      else {
+        res.status(401).json({ message: "Not authorized!" })
+      }
     })
 })
